requirements.txt
style~=1.1.0
pip~=24.0
attrs~=23.2.0
typing_extensions~=4.9.0
Pygments~=2.17.2
idna~=3.6
certifi~=2024.2.2
ipython~=8.20.0
requests~=2.31.0
aiohttp~=3.9.3
urllib3~=2.2.0
psycopg2~=2.9.9
telegram~=0.0.1
pytz~=2023.4
python-dotenv~=1.0.1